Refactored code with SonarLint

+++ src/profile/profile_stats.cpp
@@ -22,7 +22,7 @@ void profile_stats::compute_relative_durations(const profile_stats &parent)
 {
     m_total_calls = m_relative_calls * parent.m_total_calls;
 
-    m_relative_percent = ((float)m_duration_over_calls) / parent.m_duration_per_call;
+    m_relative_percent = (float)m_duration_over_calls / (float)parent.m_duration_per_call;
     m_total_percent = m_relative_percent * parent.m_total_percent;
     for (auto &[name, pstats] : m_children)
         pstats.compute_relative_durations(*this);
@@ -31,9 +31,11 @@ void profile_stats::compute_relative_durations(const profile_stats &parent)
 void profile_stats::smooth_stats(const profile_stats &stats, const float smoothness)
 {
     m_duration_per_call =
-        (long long)((1.f - smoothness) * m_duration_per_call + smoothness * stats.duration_per_call());
-    m_duration_over_calls =
-        (long long)((1.f - smoothness) * m_duration_over_calls + smoothness * stats.duration_over_calls());
+        (long long)((1.f - smoothness) * (float)m_duration_per_call + smoothness * (float)stats.duration_per_call());
+
+    m_duration_over_calls = (long long)((1.f - smoothness) * (float)m_duration_over_calls +
+                                        smoothness * (float)stats.duration_over_calls());
+
     m_relative_percent = (1.f - smoothness) * m_relative_percent + smoothness * stats.relative_percent();
     m_total_percent = (1.f - smoothness) * m_total_percent + smoothness * stats.total_percent();
     for (auto &[name, child] : m_children)
